fixed some bugs

+++ Server.h
@@ -16,7 +16,7 @@ typedef struct Server {
 
 } Server;
 
-Server setSocketAddr(int port, uint32_t ip);
+Server setSocketAddr(uint16_t port, uint32_t ip);
 int activate_server(sockaddr_in s_addr);
 void* client_handler(void* socket_desc);
 void communicate_with_client(int server_socket);
